+++ packages/gatsby-theme-basic/src/theme.js
@@ -1,3 +1,14 @@
-export default theme;
+import "./theme/colors.css";
+import "./theme/Card.props.css";
+import { Image } from "@whitespace/gatsby-theme-wordpress-basic/src/components";
 
-var theme = {};
+export default {
+  default: {
+    components: {
+      Image,
+    },
+  },
+  cardMedia: {
+    aspectRatio: 16 / 9,
+  },
+};