+++ packages/gatsby-theme-intranet/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.19.8](https://github.com/municipio-se/gatsby-packages/compare/@municipio/gatsby-theme-intranet@0.19.7...@municipio/gatsby-theme-intranet@0.19.8) (2021-11-03)
+
+**Note:** Version bump only for package @municipio/gatsby-theme-intranet
+
+
+
+
+
+### [0.19.7](https://github.com/municipio-se/gatsby-packages/compare/@municipio/gatsby-theme-intranet@0.19.6...@municipio/gatsby-theme-intranet@0.19.7) (2021-11-03)
+
+
+### Features
+
+* allow user to reorder tools in sidebar ([40a40cd](https://github.com/municipio-se/gatsby-packages/commit/40a40cd71853f8adcf9899dc2cd3938f83bd80e1))
+
+
+
+### [0.19.6](https://github.com/municipio-se/gatsby-packages/compare/@municipio/gatsby-theme-intranet@0.19.5...@municipio/gatsby-theme-intranet@0.19.6) (2021-10-27)
+
+**Note:** Version bump only for package @municipio/gatsby-theme-intranet
+
+
+
+
+
 ### [0.19.5](https://github.com/municipio-se/gatsby-packages/compare/@municipio/gatsby-theme-intranet@0.19.4...@municipio/gatsby-theme-intranet@0.19.5) (2021-10-25)